Refactoring forecast data reading
