+++ Exec/DevTests/Hurricane/GNUmakefile
@@ -14,7 +14,7 @@ USE_GPROF     = FALSE
 USE_MPI  = TRUE
 USE_OMP  = FALSE
 
-USE_CUDA = TRUE
+USE_CUDA = FALSE
 USE_HIP  = FALSE
 USE_SYCL = FALSE
 
@@ -22,7 +22,7 @@ USE_SYCL = FALSE
 DEBUG = FALSE
 
 TEST = FALSE
-USE_ASSERTION = TRUE
+USE_ASSERTION = FALSE
 
 # GNU Make
 Bpack := ./Make.package
